# Big optimization of RGBA graphics loading with fpimage (300% to 1000%)