Recently I was looking for code to easily load 8bit per channel graphics and looked at the CGE code. I found simple way to optimize RGBA graphics loading by using `FData` directly. I know that CGE can use libpng and this change can have low priority. But gain is so big  - from 300% to 1000% speed up :)

On my Ryzen 3700X loading 1920x1080 PNG file shortened from 6 ms to 2 ms. On weaker devices like android phone it's from 154 ms to 15 ms.  

On Android also smaller images have big improvement (900 x 100 from 11ms to below 1 ms, 230 x 205 from 4 ms to below 1ms ). 

Currently I only changed the loading of RGBA images, but it is possible to make fixes for RGB as well, but we would have to be able to load graphics into TFPCompactImgRGB8Bit (I think this is a bigger task for another PR). Maybe by adding more `TInternalCastleFpImage` versions.

Unfortunately, so far I have not found a way to make FPImage load graphics into `TFPCompactImgRGB8Bit` or `TFPCompactImgRGBA8Bit` depending on the file loaded. Anyone know how to do that? 

But in many cases we have information whether the file will be RGB or RGBA. For example when jpeg is loaded we can use `TFPCompactImgRGB8Bit` or from `AllowedImageClasses` parameter.

BTW. is there easy way to disable libpng on linux? 
I do something like this to test on desktop:
```
function CastlePngInitialized: boolean;
begin
  Result := false;
  Exit;
  Result := PngLibrary <> nil;
end;
```

I didn't measure how close this change brings us to libpng but I think we should make pascal solution best as we can.

Also adding a way to invert images when we "export/package" game can give as a performance boost. Then there will be only something like:

`Move(Source, Dest, RowSize * Height);` 

Without any loop.