@michaliskambi 2020-09-07
Thank you! All looks good and makes sense, great!

> Unfortunately, so far I have not found a way to make FPImage load graphics into TFPCompactImgRGB8Bit or TFPCompactImgRGBA8Bit depending on the file loaded. Anyone know how to do that?
> But in many cases we have information whether the file will be RGB or RGBA. For example when jpeg is loaded we can use TFPCompactImgRGB8Bit or from AllowedImageClasses parameter.

I don't know how to do that. Ideally FpImage would figure out by itself the optimal format, based on image header.

Note that, as usual with optimizations, I would say that they only make sense if they really cover a lot of practical use-cases :) So

- Optimizing in case of `AllowedImageClasses` doesn't contain RGBA --- this is not very important, as in most cases (when used internally in CGE) all formats are allowed at loading, as we want to allow alpha (if present in the image). 

> BTW. is there easy way to disable libpng on linux?

You can compile with `CASTLE_DISABLE_LIBPNG` symbol defined (in the project, or by CGE build tool `--compiler-option=-dCASTLE_DISABLE_LIBPNG`).

> Also adding a way to invert images when we "export/package" game can give as a performance boost. Then there will be only something like:

It is possible that some day image loaders will get `Options: TLoadImageOptions` passed from `LoadEncodedImage`. Then indeed it could be loaded using one `Move` in case of RGBA (probably not in case of other formats, as there may be padding to make row size multiple by 4; at least PNG does it, not sure if internally FpImage does this padding too).

When reading from glTF, images are usually flipped (see https://castle-engine.io/x3d_implementation_texturing_extensions.php#section_flip_vertically for description , that is because glTF has flipped interpretation of texture coordinates). So this may indeed be worthwhile at some point.